Conflicts:
	ISP392_Group1/nbproject/build-impl.xml
	ISP392_Group1/nbproject/genfiles.properties
	ISP392_Group1/nbproject/project.properties
	ISP392_Group1/nbproject/project.xml
	ISP392_Group1/src/java/category/Shop.java
	ISP392_Group1/src/java/category/ShopFilter.java
	ISP392_Group1/src/java/dal/DAO.java
	ISP392_Group1/src/java/dal/DBContext.java
	ISP392_Group1/src/java/model/Categories.java
	ISP392_Group1/src/java/model/Product.java
	ISP392_Group1/web/WEB-INF/web.xml
	ISP392_Group1/web/home.jsp
	ISP392_Group1/web/styles/shop_responsive.css
	ISP392_Group1/web/styles/shop_styles.css

+++ ISP392_Group1/src/java/model/Product.java
@@ -34,7 +34,7 @@ public class Product {
         this.discount = discount;
     }
     
-    public Product(String id, String name, double price, String image, int quantity, String description, int discount, Categories category) {
+    public Product(String id, String name, double price, String image, int quantity, String description, double discount, Categories category) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -45,7 +45,7 @@ public class Product {
         this.category = category;
     }
     
-    public Product(String id, String name, double price, String image, int quantity, String description, int discount, String status, Categories category) {
+    public Product(String id, String name, double price, String image, int quantity, String description, double discount, String status, Categories category) {
         this.id = id;
         this.name = name;
         this.price = price;
